Add GetDomains() reading models_config.yaml
It implied starting tfserving using this config file (see README.md).
We don't store the domain list we read from the config file. Indeed, it is possible to change its content on the fly and the tfserving can reload models accordingly.
TODO: improve way to get absolute path to the config file.